Move OutputParsersUtils from common to agent
OutputParsersUtils::clean() is used exclusively by Linux agent
parsers (DmesgParser, LsblkOutputParser, SmartCtlOutputParser).
Move it out of the shared common library into the agent directory
where it belongs.
