+++ src/agent/linux/DmesgParser.cpp
@@ -1,7 +1,7 @@
 
 #include <regex>
 
-#include "common/OutputParsersUtils.h"
+#include "agent/OutputParsersUtils.h"
 #include "DmesgParser.h"
 #include "IPartitionsManager.h"
