+++ src/agent/linux/LsblkOutputParser.cpp
@@ -4,7 +4,7 @@
 #include <sstream>
 #include <cassert>
 
-#include "common/OutputParsersUtils.h"
+#include "agent/OutputParsersUtils.h"
 #include "LsblkOutputParser.h"
