+++ src/agent/linux/SmartCtlOutputParser.cpp
@@ -2,7 +2,7 @@
 #include <sstream>
 #include <vector>
 
-#include "common/OutputParsersUtils.h"
+#include "agent/OutputParsersUtils.h"
 #include "SmartCtlOutputParser.h"
